fix(brk): set current p register as well as stack
On BRK, the `P` ("status") register should have the `B` flag set and the
resulting `P` value is pushed to the stack. In my implementation, the
correct value was sent to the stack, but the CPU wasnt left with the
same `P` register value. This fix sets the correct value _and_ sends it
to the stack.