add concept of IO address subscribers
This can be used as if specific hardware was attached to specific
memory addresses. For example, a screen's output may be read from
location 0xFFF0 so when a program outputs a value onto that location,
the subscriber to 0xFFF0 would be notified that a new value is at that
location.